move mmap_dense_vectors
