+++ lib/segment/src/vector_storage/dense/mod.rs
@@ -1,1 +1,2 @@
 pub mod memmap_dense_vector_storage;
+pub mod mmap_dense_vectors;
